requirements-dev.txt
# Main dependencies
-r requirements.txt

# Development dependencies
isort
pycodestyle
pylint-django
